+++ test/instruction_test.py
@@ -92,7 +92,7 @@ class TestBaseInstruction(object):
     def test_informatics():
         p = Protocol()
         cont1 = p.ref("cont1", None, "6-flat", discard=True)
-        compd = Compound("InChI=1S/CH4/h1H4")
+        compd = Compound("Daylight Canonical SMILES", "CN1C=NC2=C1C(=O)N(C(=O)N2C)C")
         example_data = {
             "wells": [cont1.well(0), cont1.well(1)],
             "some_int": 1,
@@ -102,7 +102,7 @@ class TestBaseInstruction(object):
             op="test_instruction", data=example_data, informatics=example_informatics
         )
         assert isinstance(instr.informatics[0], AttachCompounds)
-        assert instr.informatics[0].compounds[0].InChI == "InChI=1S/CH4/h1H4"
+        assert instr.informatics[0].compounds[0].value == "CN1C=NC2=C1C(=O)N(C(=O)N2C)C"
 
         instr_multi = Instruction(
             op="test_instruction",
@@ -137,7 +137,7 @@ class TestBaseInstruction(object):
     def test_info_wells_checker(self):
         p = Protocol()
         cont1 = p.ref("cont1", None, "6-flat", discard=True)
-        compd = Compound("InChI=1S/CH4/h1H4")
+        compd = Compound("Daylight Canonical SMILES", "CN1C=NC2=C1C(=O)N(C(=O)N2C)C")
         example_data = {"objects": [cont1.well(0), cont1.well(1)], "some_int": 1}
         instr = Instruction(
             op="test_instruction",