+++ docs/changelog.rst
@@ -2,6 +2,7 @@
 Changelog
 =========
 
+* :feature:`297` Add support for SMILES to Compound class
 * :bug:`295` Remove low-evaporation cover type from FALCON96UBOTTOM
 
 * :release:`7.6.1 <2021-03-04>`